Addin typecript to the middleware function
